requirements.txt
Flask
Werkzeug
PyMuPDF
python-docx
pytesseract
Pillow
langdetect
pinecone-client
sentence-transformers